Move tests into appropriate directory structure.
